+++ src/main/java/suite/trade/Strategy.java
@@ -2,7 +2,7 @@ package suite.trade;
 
 public interface Strategy {
 
-	public GetBuySell analyze(double prices[]);
+	public GetBuySell analyze(float prices[]);
 
 	// 1 = buy, 0 = no change, -1 = sell
 	public interface GetBuySell {
